+++ HISTORY.md
@@ -1,6 +1,12 @@
 # History
 ----
 
+## 0.1.2 / 2020-12-13
+
+- add util function to map popup window to/from an element
+- add initPopupInnerRect and initPopupOuterRect
+
+
 ## 0.1.0 / 2020-12-10
 
 - remove features property, change to left, top, width, height